pool gql statements and hooks decomposed
